+++ pr.md
@@ -43,6 +43,10 @@
 
 <!-- 과제에 대한 회고를 작성해주세요 -->
 
+1. displayPrice
+2. usesearchparams
+3. 액션함수 순수함수 분리
+
 ### 과제를 하면서 내가 제일 신경 쓴 부분은 무엇인가요?
 
 ### 과제를 다시 해보면 더 잘 할 수 있었겠다 아쉬운 점이 있다면 무엇인가요?
