+++ src/basic/App.tsx
@@ -1,20 +1,14 @@
 import { useCallback, useState } from "react";
 import { Coupon } from "../types";
 
-import useFilterSearchParams from "../hooks/useFilterSearchParams";
 import { AdminPage } from "./pages/AdminPage";
 import { ShopPage } from "./pages/ShopPage";
 
+import { Header } from "./components/ui/Header";
+import { NotificationList } from "./components/ui/NotificationList";
 import { useCart } from "./hooks/useCart";
 import { useCoupons } from "./hooks/useCoupons";
 import { useProducts } from "./hooks/useProducts";
-import {
-  calculateCartTotal,
-  calculateItemTotal,
-  getCartItemCount,
-} from "./utils/cartUtils";
-import { NotificationList } from "./components/ui/NotificationList";
-import { Header } from "./components/ui/Header";
 
 export interface Notification {
   id: string;
@@ -34,10 +28,10 @@ const App = () => {
   const { coupons, addCoupon, removeCoupon } = useCoupons();
 
   const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
-  const [isAdmin, setIsAdmin] = useState(false);
-  const [notifications, setNotifications] = useState<Notification[]>([]);
-  const { filterSearchParams } = useFilterSearchParams();
 
+  const [isAdmin, setIsAdmin] = useState(false);
+
+  const [notifications, setNotifications] = useState<Notification[]>([]);
   const addNotification = useCallback(
     (message: string, type: "error" | "success" | "warning" = "success") => {
       const id = Date.now().toString();
@@ -50,16 +44,8 @@ const App = () => {
     []
   );
 
-  const {
-    cart,
-    addToCart,
-    updateCartItemQuantity,
-    removeFromCart,
-    clearCart,
-    getProductRemainingStock,
-  } = useCart();
-
-  const totalItemCount = getCartItemCount(cart);
+  const { cart, addToCart, updateCartItemQuantity, removeFromCart, clearCart } =
+    useCart();
 
   return (
     <div className="min-h-screen bg-gray-50">
@@ -67,12 +53,7 @@ const App = () => {
         notifications={notifications}
         setNotifications={setNotifications}
       />
-      <Header
-        isAdmin={isAdmin}
-        setIsAdmin={setIsAdmin}
-        cart={cart}
-        totalItemCount={totalItemCount}
-      />
+      <Header isAdmin={isAdmin} setIsAdmin={setIsAdmin} cart={cart} />
 
       <main className="max-w-7xl mx-auto px-4 py-8">
         {isAdmin ? (
@@ -90,9 +71,6 @@ const App = () => {
           <ShopPage
             products={products}
             filteredProducts={filteredProducts}
-            filterSearchParams={{
-              searchTerm: filterSearchParams.searchTerm ?? "",
-            }}
             cart={cart}
             addToCart={addToCart}
             updateCartItemQuantity={updateCartItemQuantity}
@@ -101,10 +79,7 @@ const App = () => {
             selectedCoupon={selectedCoupon}
             setSelectedCoupon={setSelectedCoupon}
             addNotification={addNotification}
-            calculateCartTotal={() => calculateCartTotal(cart, selectedCoupon)}
-            calculateItemTotal={(item) => calculateItemTotal(item, cart)}
             coupons={coupons}
-            getProductRemainingStock={getProductRemainingStock}
           />
         )}
       </main>
