+++ src/basic/pages/ShopPage.tsx
@@ -1,13 +1,17 @@
 import { useCallback } from "react";
 import { CartItem, Coupon, Product } from "../../types";
-import { displayPrice } from "../utils/formatters";
-import { getRefinedProduct } from "../utils/productUtils";
-import { getRemainingStock } from "../utils/cartUtils";
+
+import { ProductList } from "../components/product/ProductList";
+import {
+  calculateCartTotal,
+  calculateItemTotal,
+  getRemainingStock,
+} from "../utils/cartUtils";
 
 interface ShopPageProps {
   products: Product[];
   filteredProducts: Product[];
-  filterSearchParams: { searchTerm?: string };
+
   cart: CartItem[];
   addToCart: (product: Product, quantity?: number) => void;
   updateCartItemQuantity: (productId: string, quantity: number) => void;
@@ -19,19 +23,12 @@ interface ShopPageProps {
     message: string,
     type: "error" | "success" | "warning"
   ) => void;
-  calculateCartTotal: () => {
-    totalBeforeDiscount: number;
-    totalAfterDiscount: number;
-  };
-  calculateItemTotal: (item: CartItem) => number;
   coupons: Coupon[];
-  getProductRemainingStock: (product: Product) => number;
 }
 
 export function ShopPage({
   products,
   filteredProducts,
-  filterSearchParams,
   cart,
   addToCart,
   updateCartItemQuantity,
@@ -40,25 +37,8 @@ export function ShopPage({
   selectedCoupon,
   setSelectedCoupon,
   addNotification,
-  calculateCartTotal,
-  calculateItemTotal,
   coupons,
 }: ShopPageProps) {
-  // 장바구니에 상품 추가 (props로 받은 함수 사용)
-  const handleAddToCart = useCallback(
-    (product: Product) => {
-      if (!getRemainingStock(product, cart)) {
-        addNotification("재고가 부족합니다", "error");
-        return;
-      }
-
-      addToCart(product, 1);
-
-      addNotification("장바구니에 담았습니다", "success");
-    },
-    [cart, addToCart, addNotification]
-  );
-
   // 장바구니 상품 수량 변경 (props로 받은 함수 사용)
   const handleUpdateQuantity = useCallback(
     (product: Product, newQuantity: number) => {
@@ -85,7 +65,10 @@ export function ShopPage({
   // 쿠폰 적용
   const applyCoupon = useCallback(
     (coupon: Coupon) => {
-      const currentTotal = calculateCartTotal().totalBeforeDiscount;
+      const currentTotal = calculateCartTotal(
+        cart,
+        selectedCoupon
+      ).totalBeforeDiscount;
 
       if (currentTotal < 10000 && coupon.discountType === "percentage") {
         addNotification(
@@ -98,7 +81,7 @@ export function ShopPage({
       setSelectedCoupon(coupon);
       addNotification("쿠폰이 적용되었습니다.", "success");
     },
-    [addNotification, calculateCartTotal, setSelectedCoupon]
+    [addNotification, cart, selectedCoupon, setSelectedCoupon]
   );
 
   // 주문 완료
@@ -112,128 +95,19 @@ export function ShopPage({
     setSelectedCoupon(null);
   }, [addNotification, clearCart, setSelectedCoupon]);
 
-  const totals = calculateCartTotal();
+  const totals = calculateCartTotal(cart, selectedCoupon);
 
   return (
     <div className="grid grid-cols-1 lg:grid-cols-4 gap-6">
       <div className="lg:col-span-3">
         {/* 상품 목록 */}
-        <section>
-          <div className="mb-6 flex justify-between items-center">
-            <h2 className="text-2xl font-semibold text-gray-800">전체 상품</h2>
-            <div className="text-sm text-gray-600">
-              총 {products.length}개 상품
-            </div>
-          </div>
-          {filteredProducts.length === 0 ? (
-            <div className="text-center py-12">
-              <p className="text-gray-500">
-                "{filterSearchParams.searchTerm ?? ""}"에 대한 검색 결과가
-                없습니다.
-              </p>
-            </div>
-          ) : (
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-              {filteredProducts.map((product) => {
-                const refinedProduct = getRefinedProduct(product, cart);
-
-                return (
-                  <div
-                    key={product.id}
-                    className="bg-white rounded-lg border border-gray-200 overflow-hidden hover:shadow-lg transition-shadow"
-                  >
-                    {/* 상품 이미지 영역 (placeholder) */}
-                    <div className="relative">
-                      <div className="aspect-square bg-gray-100 flex items-center justify-center">
-                        <svg
-                          className="w-24 h-24 text-gray-300"
-                          fill="none"
-                          stroke="currentColor"
-                          viewBox="0 0 24 24"
-                        >
-                          <path
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                            strokeWidth={1}
-                            d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z"
-                          />
-                        </svg>
-                      </div>
-                      {product.isRecommended && (
-                        <span className="absolute top-2 right-2 bg-red-500 text-white text-xs px-2 py-1 rounded">
-                          BEST
-                        </span>
-                      )}
-                      {product.discounts.length > 0 && (
-                        <span className="absolute top-2 left-2 bg-orange-500 text-white text-xs px-2 py-1 rounded">
-                          ~
-                          {Math.max(...product.discounts.map((d) => d.rate)) *
-                            100}
-                          %
-                        </span>
-                      )}
-                    </div>
-
-                    {/* 상품 정보 */}
-                    <div className="p-4">
-                      <h3 className="font-medium text-gray-900 mb-1">
-                        {product.name}
-                      </h3>
-                      {product.description && (
-                        <p className="text-sm text-gray-500 mb-2 line-clamp-2">
-                          {product.description}
-                        </p>
-                      )}
-
-                      {/* 가격 정보 */}
-                      <div className="mb-3">
-                        <p className="text-lg font-bold text-gray-900">
-                          {displayPrice(refinedProduct, {
-                            prefix: "₩",
-                          })}
-                        </p>
-                        {product.discounts.length > 0 && (
-                          <p className="text-xs text-gray-500">
-                            {product.discounts[0].quantity}개 이상 구매시 할인{" "}
-                            {product.discounts[0].rate * 100}%
-                          </p>
-                        )}
-                      </div>
-
-                      {/* 재고 상태 */}
-                      <div className="mb-3">
-                        {refinedProduct.stock <= 5 &&
-                          refinedProduct.stock > 0 && (
-                            <p className="text-xs text-red-600 font-medium">
-                              품절임박! {refinedProduct.stock}개 남음
-                            </p>
-                          )}
-                        {refinedProduct.stock > 5 && (
-                          <p className="text-xs text-gray-500">
-                            재고 {refinedProduct.stock}개
-                          </p>
-                        )}
-                      </div>
-
-                      {/* 장바구니 버튼 */}
-                      <button
-                        onClick={() => handleAddToCart(product)}
-                        disabled={refinedProduct.stock <= 0}
-                        className={`w-full py-2 px-4 rounded-md font-medium transition-colors ${
-                          refinedProduct.stock <= 0
-                            ? "bg-gray-100 text-gray-400 cursor-not-allowed"
-                            : "bg-gray-900 text-white hover:bg-gray-800"
-                        }`}
-                      >
-                        {refinedProduct.stock <= 0 ? "품절" : "장바구니 담기"}
-                      </button>
-                    </div>
-                  </div>
-                );
-              })}
-            </div>
-          )}
-        </section>
+        <ProductList
+          products={products}
+          filteredProducts={filteredProducts}
+          cart={cart}
+          addToCart={addToCart}
+          addNotification={addNotification}
+        />
       </div>
 
       <div className="lg:col-span-1">
@@ -275,7 +149,7 @@ export function ShopPage({
             ) : (
               <div className="space-y-3">
                 {cart.map((item) => {
-                  const itemTotal = calculateItemTotal(item);
+                  const itemTotal = calculateItemTotal(item, cart);
                   const originalPrice = item.product.price * item.quantity;
                   const hasDiscount = itemTotal < originalPrice;
                   const discountRate = hasDiscount
